AiModule/requirements.txt
faiss-cpu==1.10.0
google-generativeai==0.8.4
grpcio==1.71.0
grpcio-tools==1.71.0
numpy==2.2.4
sentence-transformers==3.4.1
protobuf==5.29.3
huggingface_hub[hf_xet]>=0.15.0
tiktoken